pyproject.toml
[project]
name = "gemini-tts-bot"
version = "0.1.0"
description = "Telegram Bot for Gemini TTS"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "google-genai>=1.0.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gemini_tts_bot"]

[project.scripts]
gemini-tts-bot = "gemini_tts_bot.main:main"
